pyproject.toml
[project]
name = "mavi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "encord>=0.1.159",
    "encord-agents>=0.1.11",
    "inquirerpy>=0.3.4",
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "lxml>=5.3.2",
    "modal>=0.73.149",
    "pydantic>=2.11.2",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "selenium>=4.31.0",
    "webdriver-manager>=4.0.2",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "ipython>=9.1.0",
    "mypy>=1.15.0",
    "pyfzf>=0.3.1",
    "ruff>=0.11.4",
    "types-requests>=2.32.0.20250328",
]
